+++ pkg/operator/operands/common/common.go
@@ -31,8 +31,13 @@ var KaiServicesForServiceMonitor = []struct {
 	Name     string
 	Port     string
 	JobLabel string
+
+	LabelSelector map[string]string // optional, if not provided, "app": NAME will be used
+	Namespaces    []string          // optional, if not provided, the service will be monitored in the namespace of the KAI config
 }{
-	{"queue-controller", "metrics", "queue-controller"},
+	{"queue-controller", "metrics", "queue-controller", nil, nil},
+	{"kube-state-metrics", "http", "kube-state-metrics",
+		map[string]string{"app.kubernetes.io/name": "kube-state-metrics"}, []string{"monitoring", "default"}},
 }
 
 func AllControllersAvailable(